Tidied progress bar. Less frequent updates and no \r better anyway, plus hopefully will avoid RStudio stack imbalance, #2481

+++ src/fread.h
@@ -343,7 +343,7 @@ void freeThreadContext(ThreadLocalFreadParsingContext *ctx);
 /**
  * Progress-reporting function.
  */
-void progress(double percent/*[0,1]*/, double ETA/*secs*/);
+void progress(int percent/*[0,100]*/, int ETA/*secs*/);
 
 
 bool freadCleanup(void);
